Hacker_News/requirements.txt
alembic==1.12.1
astroid==3.0.1
Authlib==1.2.1
Babel==2.13.1
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
coverage==7.3.2
cryptography==41.0.5
dill==0.3.7
exceptiongroup==1.2.0
Flask==3.0.0
Flask-BabelEx==0.9.4
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.0
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.2
passlib==1.7.4
platformdirs==4.0.0
pluggy==1.3.0
pycparser==2.21
pylint==3.0.2
pytest==7.4.3
python-dotenv==1.0.0
requests==2.31.0
speaklater==1.3
SQLAlchemy==2.0.22
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.0
WTForms==3.1.1
